Enviroment Safety Checks
Antistasi Dev Deploy now creates the output folder if it doesn't exist.

If your Arma 3 Profile Name doesn't exist or is missing (ie. A3 not installed), and you have not selected an output overried: It will now export to PackagedMissions in the same Directory.

(Requires Use of Configurator) Option added to open the output folder in Windows Explorer.